Renaming file containing widget helpers for clarity.
