+++ graylog2-web-interface/src/views/components/Query.jsx
@@ -5,7 +5,7 @@ import { Col } from 'components/graylog';
 
 import { Spinner } from 'components/common';
 
-import { widgetDefinition } from 'views/logic/Widget';
+import { widgetDefinition } from 'views/logic/Widgets';
 import WidgetGrid from 'views/components/WidgetGrid';
 import WidgetPosition from 'views/logic/widgets/WidgetPosition';
 import { CurrentViewStateActions } from 'views/stores/CurrentViewStateStore';
