+++ graylog2-web-interface/src/views/components/dashboard/DashboardWidget.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import { BootstrapModalConfirm } from 'components/bootstrap';
-import { widgetDefinition } from 'views/logic/Widget';
+import { widgetDefinition } from 'views/logic/Widgets';
 
 import WidgetFrame from '../widgets/WidgetFrame';
 import DashboardWidgetHeader from './DashboardWidgetHeader';
