+++ graylog2-web-interface/src/views/logic/Widgets.js
